Refactor comment and reply create tests
- refactor comment create test
- refactor reply create test
- fix comment controller to return `status 201` on create
- drop old comment and reply create tests

See #241
